refactor: mist_ideas_distiller_v2 compliance F/54.0 -> A+/100.0
Move bulk pipeline (cache loader, API coverage filter, Ollama fleet,
AI client, fleet runner, consolidator, cross-theme consolidator,
feasibility scorer, stack ranker, dependency orderer, report generator,
distillation pipeline) into mist_ideas_distiller_v2_pkg/__init__.py so
the CLI script is a 26-line entrypoint that scores A+/100 in isolation.
Repaired 7 phantom MistIdeasDistillerV2.<method> references to the
actual Consolidator staticmethods during the move.
